pyproject.toml
[tool.poetry]
name = "stocker-python"
version = "0.1.0"
description = ""
authors = ["Jared Head <jaredheadd@gmail.com>"]
readme = "README.md"
packages = [{include = "stocker_python"}]

[tool.poetry.dependencies]
python = "^3.11"
python-dotenv = "^1.0.0"
beautifulsoup4 = "^4.12.2"
requests = "^2.31.0"
mysqlclient = "^2.2.0"
Flask = "^3.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
